tweak

+++ GameEngine/GameLogic/Scene/Trap/BearTrap/BearTrap.cpp
@@ -53,7 +53,7 @@ void BearTrap::Update(double deltaTime)
 	//this->isBeingReloaded = false; //Needs to be reset every frame, in case the player releases the button
 	renderInfo.normalizedReloadVal = -1;
 	if (this->currReloadTime > 0)
-		currReloadTime -=  deltaTime/4;
+		currReloadTime -=  (float)(deltaTime/4);
 	
 	if (this->GetState()->GetTrapState() == TrapState::TRAP_ACTIVE_STATE)
 	{
